Restructure search objects around seller versions
