+++ app/views/sellers/search/_result.html.erb
@@ -1,5 +1,5 @@
 <li class="result">
-  <h3><%= link_to seller.name, sellers_profile_path(seller) %></h3>
-  <div class="tags"><%= seller_profile_tags(seller) %></div>
-  <p><%= seller.summary %></p>
+  <h3><%= link_to seller_version.name, sellers_profile_path(seller_version) %></h3>
+  <div class="tags"><%= seller_profile_tags(seller_version) %></div>
+  <p><%= seller_version.summary %></p>
 </li>
